Fix GPU-CPU sync bottleneck and add training profiler
- Replace .item() calls inside training loop with tensor accumulation
  to avoid GPU-CPU synchronization every batch
- Add SimpleProfiler class for timing analysis of training phases
- Add --profile CLI flag to enable profiler output

The .item() calls were causing a sync every batch, significantly
slowing down training. Now accumulators stay as tensors and only
convert to scalars at epoch end.

+++ lib/engine_mae.py
@@ -216,7 +216,7 @@ def run_epoch_mae(model, optimizer, device, root, tree,
                 # Already-invalid sensors have time == sentinel_value and are NOT in mask
                 already_invalid = (x_in[:, :, 1] == sentinel_value)  # (B, N)
                 total_valid_sensors += (~already_invalid).sum()
-                total_randomly_masked += mask.sum()
+                total_randomly_masked += mask.sum().long()
 
                 profiler.start("loss_compute")
                 # 2. Gather Truth Targets
@@ -668,7 +668,7 @@ def run_eval_mae(model, device, root, tree,
                     # Track actual mask ratio (no .item() to avoid GPU-CPU sync)
                     already_invalid = (x_in[:, :, 1] == sentinel_value)  # (B, N)
                     total_valid_sensors += (~already_invalid).sum()
-                    total_randomly_masked += mask.sum()
+                    total_randomly_masked += mask.sum().long()
 
                     latent_seq = model.encoder.forward_features(x_masked)
